requirements.txt
Flask
SQLAlchemy
pytest
python-dotenv
psycopg2-binary
pytest-order
pytest-ordering